Use AWS roles instead of direct user credentials
When acquiring AWS session tokens, allow user to define role that
should be assumed. This way we can use roles and attach permissions
to roles instead of directly attaching them to user accouts.

This allows us to limit permissions applied when running infrastructure
administration steps and we can avoind accidentally doing things that
should not be part of those steps.